+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### FEATURES:
 * builder/docker: add support for packer's docker builder
 * builder/file: add support for packer's file builder
+* post_processor/Checksum: add support for packer's checksum post-processor
 * post_processor/DockerImport: add support for packer's docker import post-processor
 * post_processor/DockerPush: add support for packer's docker push post-processor
 * post_processor/DockerSave: add support for packer's docker save post-processor